+++ modules/common/src/main/scala/higherkindness/mu/rpc/protocol/protocol.scala
@@ -33,11 +33,14 @@ sealed abstract class CompressionType extends Product with Serializable
 case object Identity                  extends CompressionType
 case object Gzip                      extends CompressionType
 
+sealed abstract class MethodNameStyle extends Product with Serializable
+case object Unchanged                 extends MethodNameStyle
+case object Capitalize                extends MethodNameStyle
+
 class message                          extends StaticAnnotation
 class option(name: String, value: Any) extends StaticAnnotation
 class outputPackage(value: String)     extends StaticAnnotation
 class outputName(value: String)        extends StaticAnnotation
-class http                             extends StaticAnnotation
 
 @message
 object Empty