+++ modules/internal/monix/src/main/scala/higherkindness/mu/client/monixCalls.scala
@@ -22,24 +22,13 @@ import cats.effect.Async
 import io.grpc.stub.{ClientCalls, StreamObserver}
 import io.grpc.{CallOptions, Channel, MethodDescriptor}
 import monix.reactive.Observable
-import higherkindness.mu.rpc.internal.task._
+import monix.execution.Scheduler
 import monix.execution.rstreams.Subscription
 import org.reactivestreams.{Publisher, Subscriber}
-
-import scala.concurrent.ExecutionContext
 
 object monixCalls {
 
   import higherkindness.mu.rpc.internal.converters._
-
-  def unary[F[_]: Async, Req, Res](
-      request: Req,
-      descriptor: MethodDescriptor[Req, Res],
-      channel: Channel,
-      options: CallOptions)(implicit EC: ExecutionContext): F[Res] =
-    listenableFuture2Async[F, Res](
-      ClientCalls
-        .futureUnaryCall(channel.newCall(descriptor, options), request))
 
   def serverStreaming[Req, Res](
       request: Req,
@@ -53,7 +42,7 @@ object monixCalls {
       input: Observable[Req],
       descriptor: MethodDescriptor[Req, Res],
       channel: Channel,
-      options: CallOptions)(implicit EC: ExecutionContext): F[Res] =
+      options: CallOptions)(implicit S: Scheduler): F[Res] =
     input
       .liftByOperator(
         StreamObserver2MonixOperator(